backend/requirements.txt
fastapi
uvicorn
pyserial
sqlmodel
qrcode
websockets
aiofiles
